# openapi.json
{
  "components": {
    "schemas": {
...
      "Body_uploadDocument": {
        "properties": {
          "file": {
            "format": "binary",
            "title": "File",
            "type": "string"
          },
          "projectId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Projectid"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_uploadDocument",
        "type": "object"
      },
...
      "Chat": {
        "additionalProperties": false,
        "example": {
          "createdAt": "2024-01-15T10:00:00Z",
          "id": "c4",
          "projectId": "work",
          "title": "Q1 metrics summary",
          "updatedAt": "2024-01-15T10:00:00Z"
        },
        "properties": {
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "projectId": {
            "title": "Projectid",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "projectId",
          "createdAt",
          "updatedAt"
        ],
        "title": "Chat",
        "type": "object"
      },
      "ChatSearchResult": {
        "additionalProperties": false,
        "example": {
          "createdAt": "2024-01-15T10:00:00Z",
          "id": "c4",
          "projectId": "work",
          "projectName": "Work",
          "title": "Q1 metrics summary",
          "updatedAt": "2024-01-15T10:00:00Z"
        },
        "properties": {
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "projectId": {
            "title": "Projectid",
            "type": "string"
          },
          "projectName": {
            "title": "Projectname",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "projectId",
          "projectName",
          "createdAt",
          "updatedAt"
        ],
        "title": "ChatSearchResult",
        "type": "object"
      },
...
      "CreateChatRequest": {
        "additionalProperties": false,
        "example": {
          "title": "New chat"
        },
        "properties": {
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title"
        ],
        "title": "CreateChatRequest",
        "type": "object"
      },
...
      "Document": {
        "additionalProperties": false,
        "example": {
          "addedAt": "2024-01-15T10:00:00Z",
          "id": "doc-1",
          "name": "Q1 Financial Report.pdf",
          "size": 2457600,
          "status": "ingested",
          "type": "application/pdf"
        },
        "properties": {
          "addedAt": {
            "format": "date-time",
            "title": "Addedat",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "projectId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Projectid"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/DocumentStatus"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "size",
          "type",
          "status",
          "addedAt"
        ],
        "title": "Document",
        "type": "object"
      },
      "DocumentStatus": {
        "enum": [
          "pending",
          "ingested",
          "error"
        ],
        "title": "DocumentStatus",
        "type": "string"
      },
      "DocumentStatusResponse": {
        "additionalProperties": false,
        "example": {
          "id": "doc-1",
          "status": "ingested"
        },
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/DocumentStatus"
          }
        },
        "required": [
          "id",
          "status"
        ],
        "title": "DocumentStatusResponse",
        "type": "object"
      },
...
      "PaginatedChatSearchResults": {
        "additionalProperties": false,
        "properties": {
          "hasMore": {
            "default": false,
            "title": "Hasmore",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ChatSearchResult"
            },
            "title": "Items",
            "type": "array"
          },
          "nextCursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextcursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "PaginatedChatSearchResults",
        "type": "object"
      },
      "PaginatedChats": {
        "additionalProperties": false,
        "properties": {
          "hasMore": {
            "default": false,
            "title": "Hasmore",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/Chat"
            },
            "title": "Items",
            "type": "array"
          },
          "nextCursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextcursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "PaginatedChats",
        "type": "object"
      },
      "PaginatedDocuments": {
        "additionalProperties": false,
        "properties": {
          "hasMore": {
            "default": false,
            "title": "Hasmore",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/Document"
            },
            "title": "Items",
            "type": "array"
          },
          "nextCursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextcursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "PaginatedDocuments",
        "type": "object"
      },
...
      "UpdateChatRequest": {
        "additionalProperties": false,
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "UpdateChatRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Databricks Apps FastAPI Starter API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/chats/recent": {
      "get": {
        "operationId": "getRecentChats",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedChatSearchResults"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Recent Chats",
        "tags": [
          "chats"
        ]
      }
    },
    "/chats/search": {
      "get": {
        "operationId": "searchChats",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedChatSearchResults"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Search Chats",
        "tags": [
          "chats"
        ]
      }
    },
    "/chats/{chatId}": {
      "delete": {
        "operationId": "deleteChat",
        "parameters": [
          {
            "in": "path",
            "name": "chatId",
            "required": true,
            "schema": {
              "title": "Chatid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "summary": "Delete Chat",
        "tags": [
          "chats"
        ]
      },
      "patch": {
        "operationId": "updateChat",
        "parameters": [
          {
            "in": "path",
            "name": "chatId",
            "required": true,
            "schema": {
              "title": "Chatid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Chat"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Update Chat",
        "tags": [
          "chats"
        ]
      }
    },
...
    "/documents": {
      "get": {
        "operationId": "listDocuments",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/DocumentStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "projectId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Projectid"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedDocuments"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Documents",
        "tags": [
          "documents"
        ]
      },
      "post": {
        "operationId": "uploadDocument",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_uploadDocument"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Document"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Upload Document",
        "tags": [
          "documents"
        ]
      }
    },
    "/documents/{documentId}": {
      "delete": {
        "operationId": "deleteDocument",
        "parameters": [
          {
            "in": "path",
            "name": "documentId",
            "required": true,
            "schema": {
              "title": "Documentid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "summary": "Delete Document",
        "tags": [
          "documents"
        ]
      }
    },
    "/documents/{documentId}/status": {
      "get": {
        "operationId": "getDocumentStatus",
        "parameters": [
          {
            "in": "path",
            "name": "documentId",
            "required": true,
            "schema": {
              "title": "Documentid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Document Status",
        "tags": [
          "documents"
        ]
      }
    },
...
    "/projects/{projectId}/chats": {
      "get": {
        "operationId": "listProjectChats",
        "parameters": [
          {
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "title": "Projectid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedChats"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Project Chats",
        "tags": [
          "chats"
        ]
      },
      "post": {
        "operationId": "createProjectChat",
        "parameters": [
          {
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "title": "Projectid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Chat"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Create Project Chat",
        "tags": [
          "chats"
        ]
      }
    },
...
  }
}